+++ resources/config.rb
@@ -7,9 +7,9 @@ property :start_on_boot, [true, false], default: true
 property :max_open_files, Integer, default: 131_072
 property :max_locked_memory, Integer, default: 82_000
 property :instance_name, String, default: VarnishCookbook::Helpers.hostname
-property :major_version, Float, equal_to: [3.0, 4.0, 4.1, 5, 5.0, 5.1, 5.2, 6.0, 6.1], default: lazy {
-  VarnishCookbook::Helpers.installed_major_version
-}
+property :major_version, Float,
+          equal_to: [3.0, 4.0, 4.1, 5, 5.0, 5.1, 5.2, 6.0, 6.1, 6.2, 6.3, 6.4, 6.5, 6.6, 7.0],
+          default: lazy { VarnishCookbook::Helpers.installed_major_version }
 
 # Daemon options
 property :listen_address, String, default: '0.0.0.0'